+++ CDJSVis/visclibs/vectors.h
@@ -4,3 +4,5 @@ int separationVector(int, double*, double*, double*, double*, int *);
 double separationMagnitude(int, double*, double*, double*, int *);
 
 double magnitude(int, double*);
+
+int eliminatePBCFlicker(int, double *, double *, double *, int *);